Refactor: Update 함수 수정
