+++ src/components/ContentEditor/index.tsx
@@ -8,7 +8,7 @@ import EditorHeader from './EditorHeader';
 // states
 import { paramState, selectedDocumentState } from 'atoms';
 // hooks
-import { useUpdateDocs } from 'hooks';
+import { useUpdateTodo } from 'hooks';
 // styles
 import '@toast-ui/editor/dist/toastui-editor.css';
 import '@toast-ui/editor/dist/i18n/ko-kr';
@@ -18,7 +18,7 @@ function ContentEditor() {
   const params = useRecoilValue(paramState);
   const document = useRecoilValue(selectedDocumentState);
   const flag = useRef(false);
-  const updator = useUpdateDocs();
+  const updator = useUpdateTodo();
   const editorRef = useMemo(() => createRef<Editor>(), []);
   let timer: NodeJS.Timeout;
 
@@ -29,7 +29,7 @@ function ContentEditor() {
     }
     timer = setTimeout(async () => {
       if (!document) return;
-      await updator(document, 'content', content, true);
+      await updator(document, 'content', content);
     }, 1000);
   };
 
@@ -37,7 +37,7 @@ function ContentEditor() {
     const content = editorRef.current?.getInstance().getMarkdown();
     setTimeout(async () => {
       if (!document) return;
-      await updator(document, 'content', content, true);
+      await updator(document, 'content', content);
     }, 100);
   };
