+++ src/hooks/index.ts
@@ -1,5 +1,5 @@
 export * from './useGetDocRef';
 export * from './useGetListDocRef';
-export * from './useUpdateDocs';
+export * from './useUpdateTodo';
 export * from './useSetDocCount';
 export * from './useDetectClickOutside';
